UPDATE:- count datatype change

+++ Model_15M/models.py
@@ -18,7 +18,7 @@ class CONFIG_15M(models.Model):
     stoploss                = models.FloatField(verbose_name='STOPLOSS',default=0)
     target                  = models.FloatField(verbose_name='TARGET',default=0)
     quantity                = models.BigIntegerField(verbose_name='QUANTITY',default=0)
-    count                   = models.BigIntegerField(verbose_name='COUNT',default=0)
+    count                   = models.FloatField(verbose_name='COUNT',default=0)
     return_price            = models.FloatField(verbose_name='RET_PRICE',default=0)
     ltp                     = models.FloatField(verbose_name='LTP',default=0)
     order_id                = models.BigIntegerField(verbose_name='ORDER ID',default=0)
